refactor mulh to support opcode circuit smooth migration
