+++ ceno_zkvm/src/instructions/riscv.rs
@@ -18,7 +18,7 @@ pub mod ecall_base;
 pub mod jump;
 pub mod logic;
 pub mod logic_imm;
-pub mod mul;
+pub mod mulh;
 pub mod shift;
 pub mod shift_imm;
 pub mod slt;
